[Frontend] Relationship HorizontalBars widget refetch (#14448)
