--- opencti-platform/opencti-front/src/private/components/common/stix_relationships/StixRelationshipsHorizontalBars.jsx
@@ -1,200 +1,0 @@
-import { graphql } from 'react-relay';
-import { QueryRenderer } from '../../../../relay/environment';
-import { useFormatter } from '../../../../components/i18n';
-import { buildFiltersAndOptionsForWidgets } from '../../../../utils/filters/filtersUtils';
-import WidgetContainer from '../../../../components/dashboard/WidgetContainer';
-import WidgetNoData from '../../../../components/dashboard/WidgetNoData';
-import WidgetHorizontalBars from '../../../../components/dashboard/WidgetHorizontalBars';
-import useDistributionGraphData from '../../../../utils/hooks/useDistributionGraphData';
-import Loader, { LoaderVariant } from '../../../../components/Loader';
-import { useState } from 'react';
-import useDashboardViz from '../../../../components/dashboard/useDashboardViz';
-import WidgetNoHostEntity from '../../../../components/dashboard/WidgetNoHostEntity';
-
-export const stixRelationshipsHorizontalBarsDistributionQuery = graphql`
-  query StixRelationshipsHorizontalBarsDistributionQuery(
-    $field: String!
-    $operation: StatsOperation!
-    $startDate: DateTime
-    $endDate: DateTime
-    $dateAttribute: String
-    $isTo: Boolean
-    $limit: Int
-    $fromOrToId: [String]
-    $elementWithTargetTypes: [String]
-    $fromId: [String]
-    $fromRole: String
-    $fromTypes: [String]
-    $toId: [String]
-    $toRole: String
-    $toTypes: [String]
-    $relationship_type: [String]
-    $confidences: [Int]
-    $search: String
-    $filters: FilterGroup
-    $dynamicFrom: FilterGroup
-    $dynamicTo: FilterGroup
-  ) {
-    stixRelationshipsDistribution(
-      field: $field
-      operation: $operation
-      startDate: $startDate
-      endDate: $endDate
-      dateAttribute: $dateAttribute
-      isTo: $isTo
-      limit: $limit
-      fromOrToId: $fromOrToId
-      elementWithTargetTypes: $elementWithTargetTypes
-      fromId: $fromId
-      fromRole: $fromRole
-      fromTypes: $fromTypes
-      toId: $toId
-      toRole: $toRole
-      toTypes: $toTypes
-      relationship_type: $relationship_type
-      confidences: $confidences
-      search: $search
-      filters: $filters
-      dynamicFrom: $dynamicFrom
-      dynamicTo: $dynamicTo
-    ) {
-      label
-      value
-      entity {
-        ... on BasicObject {
-          id
-          entity_type
-        }
-        ... on BasicRelationship {
-          id
-          entity_type
-        }
-        ... on StixObject {
-          representative {
-            main
-          }
-        }
-        ... on StixRelationship {
-          representative {
-            main
-          }
-        }
-        # use colors when available
-        ... on Label {
-          color
-        }
-        ... on MarkingDefinition {
-          x_opencti_color
-        }
-        # objects without representative
-        ... on Creator {
-          name
-        }
-        ... on Group {
-          name
-        }
-        ... on Status {
-          template {
-            name
-            color
-          }
-        }
-      }
-    }
-  }
-`;
-
-const StixRelationshipsHorizontalBars = ({
-  title,
-  variant,
-  height,
-  field,
-  isWidget = true,
-  startDate,
-  endDate,
-  dataSelection,
-  fromId,
-  relationshipType,
-  parameters = {},
-  popover,
-  host,
-}) => {
-  const { t_i18n } = useFormatter();
-  const [chart, setChart] = useState();
-  const { resolvedDataSelection, isMissingHostEntity, isPreviewMode } = useDashboardViz({
-    perspective: 'relationships',
-    dataSelection,
-    host,
-  });
-
-  const { buildWidgetProps } = useDistributionGraphData();
-  const renderContent = () => {
-    if (isMissingHostEntity) {
-      return <WidgetNoHostEntity host={host} />;
-    }
-    let selection = {};
-    let filtersAndOptions;
-    if (resolvedDataSelection) {
-      selection = resolvedDataSelection[0];
-      filtersAndOptions = buildFiltersAndOptionsForWidgets(selection.filters, { isKnowledgeRelationshipWidget: true });
-    }
-    const finalField = selection.attribute || field || 'entity_type';
-    const variables = {
-      field: finalField,
-      operation: 'count',
-      startDate,
-      endDate,
-      fromId,
-      toTypes: !isWidget ? ['Stix-Core-Object'] : null,
-      relationship_type: !isWidget ? relationshipType : null,
-      dateAttribute: selection.date_attribute ?? 'created_at',
-      limit: selection.number ?? 10,
-      filters: filtersAndOptions?.filters,
-      isTo: selection.isTo,
-      dynamicFrom: selection.dynamicFrom,
-      dynamicTo: selection.dynamicTo,
-    };
-    return (
-      <QueryRenderer
-        query={stixRelationshipsHorizontalBarsDistributionQuery}
-        variables={variables}
-        render={({ props }) => {
-          if (
-            props
-            && props.stixRelationshipsDistribution
-            && props.stixRelationshipsDistribution.length > 0
-          ) {
-            const { series, redirectionUtils } = buildWidgetProps(props.stixRelationshipsDistribution, selection, 'Number of relationships');
-            return (
-              <WidgetHorizontalBars
-                series={series}
-                distributed={parameters.distributed}
-                onMounted={setChart}
-                redirectionUtils={redirectionUtils}
-              />
-            );
-          }
-          if (props) {
-            return <WidgetNoData />;
-          }
-          return <Loader variant={LoaderVariant.inElement} />;
-        }}
-      />
-    );
-  };
-  return (
-    <WidgetContainer
-      padding="small"
-      height={height}
-      title={parameters.title ?? title ?? t_i18n('Distribution of entities')}
-      variant={variant}
-      chart={chart}
-      action={popover}
-      showPreviewTag={isPreviewMode}
-    >
-      {renderContent()}
-    </WidgetContainer>
-  );
-};
-
-export default StixRelationshipsHorizontalBars;
